feat: Refactor Address button into a reusable component.
