+++ src/components/Guilds/Header/WalletButton.tsx
@@ -1,71 +1,15 @@
-import { isDesktop } from 'react-device-detect';
-import styled from 'styled-components';
-import { DEFAULT_ETH_CHAIN_ID, getChains, injected } from 'provider/connectors';
-import { useEffect, useMemo, useState } from 'react';
+import { getChains, injected } from 'provider/connectors';
+import { useEffect, useState } from 'react';
 import { useRpcUrls } from 'provider/providerHooks';
-import { Button, IconButton } from '../common/Button';
-import useENSAvatar from '../../../hooks/Guilds/ens/useENSAvatar';
-import Avatar from '../Avatar';
-import { shortenAddress } from '../../../utils';
+import { Button } from '../common/Button';
 import { useWeb3React } from '@web3-react/core';
 import WalletModal from '../Web3Modals/WalletModal';
-import { Badge } from '../common/Badge';
-
-const IconHolder = styled.span`
-  display: flex;
-  justify-content: center;
-
-  @media only screen and (min-width: 768px) {
-    margin-right: 0.3rem;
-  }
-
-  img {
-    border-radius: 50%;
-    margin-right: 0;
-  }
-`;
-
-const AccountButton = styled(IconButton)`
-  margin-top: 0;
-  margin-bottom: 0;
-  padding: 0.3rem;
-
-  @media only screen and (min-width: 768px) {
-    padding: 0.3rem 0.5rem;
-  }
-
-  /* hover state for when having child Badge */
-  &:hover,
-  &:active {
-    ${Badge} {
-      background-color: ${({ theme }) => theme.colors.background};
-      color: ${({ theme }) => theme.colors.primary};
-    }
-  }
-`;
-
-const AddressText = styled.span`
-  margin-left: 0.2rem;
-  margin-right: 0.3rem;
-`;
+import AddressButton from '../AddressButton';
 
 const Web3Status = () => {
   const { account, chainId } = useWeb3React();
-  const { ensName, imageUrl, avatarUri } = useENSAvatar(
-    account,
-    DEFAULT_ETH_CHAIN_ID
-  );
   const [isWalletModalOpen, setIsWalletModalOpen] = useState(false);
 
-  let imageUrlToUse = useMemo(() => {
-    if (avatarUri) {
-      return (
-        imageUrl || `https://metadata.ens.domains/mainnet/avatar/${ensName}`
-      );
-    } else {
-      return null;
-    }
-  }, [imageUrl, ensName, avatarUri]);
 
   const [injectedWalletAuthorized, setInjectedWalletAuthorized] =
     useState(false);
@@ -120,17 +64,11 @@ const Web3Status = () => {
       );
     } else if (account) {
       return (
-        <AccountButton onClick={toggleWalletModal} iconLeft>
-          <IconHolder>
-            <Avatar src={imageUrlToUse} defaultSeed={account} size={24} />
-          </IconHolder>
-          {isDesktop && (
-            <AddressText>{ensName || shortenAddress(account)}</AddressText>
-          )}
-          {transactionsCounter && (
-            <Badge size="25">{transactionsCounter}</Badge>
-          )}
-        </AccountButton>
+        <AddressButton
+          address={account}
+          transactionsCounter={transactionsCounter}
+          onClick={toggleWalletModal}
+        />
       );
     } else {
       return <Button onClick={toggleWalletModal}>Connect Wallet</Button>;
